[SYSTEMML-2190] Change parser/hops block sizes from long to int
This patch refactors the parser and hop data structures to use int
instead of long block sizes for consistency the runtime and avoid
unnecessarily bloated compiler data structures.

+++ src/main/java/org/apache/sysml/hops/DataOp.java
@@ -80,7 +80,7 @@ public class DataOp extends Hop
 	 * @param colsPerBlock cols per block
 	 */
 	public DataOp(String l, DataType dt, ValueType vt, DataOpTypes dop,
-			String fname, long dim1, long dim2, long nnz, long rowsPerBlock, long colsPerBlock) {
+			String fname, long dim1, long dim2, long nnz, int rowsPerBlock, int colsPerBlock) {
 		super(l, dt, vt);
 		_dataop = dop;
 		
@@ -96,7 +96,7 @@ public class DataOp extends Hop
 	}
 
 	public DataOp(String l, DataType dt, ValueType vt, DataOpTypes dop,
-			String fname, long dim1, long dim2, long nnz, UpdateType update, long rowsPerBlock, long colsPerBlock) {
+			String fname, long dim1, long dim2, long nnz, UpdateType update, int rowsPerBlock, int colsPerBlock) {
 		this(l, dt, vt, dop, fname, dim1, dim2, nnz, rowsPerBlock, colsPerBlock);
 		setUpdateType(update);
 	}
@@ -218,7 +218,7 @@ public class DataOp extends Hop
 		_dataop = type;
 	}
 	
-	public void setOutputParams(long dim1, long dim2, long nnz, UpdateType update, long rowsPerBlock, long colsPerBlock) {
+	public void setOutputParams(long dim1, long dim2, long nnz, UpdateType update, int rowsPerBlock, int colsPerBlock) {
 		setDim1(dim1);
 		setDim2(dim2);
 		setNnz(nnz);
